+++ pycobertura/cobertura.py
@@ -17,6 +17,12 @@ from pycobertura.utils import (
     memoize,
 )
 
+try:
+    basestring
+except NameError:
+    # PY3 basestring
+    basestring = (str, bytes)
+
 
 class Line(namedtuple('Line', ['number', 'source', 'status', 'reason'])):
     """
@@ -35,26 +41,40 @@ class Cobertura(object):
     """
     An XML Cobertura parser.
     """
-    def __init__(self, xml_path, base_path=None):
+    def __init__(self, report, source=None, source_prefix=None):
         """
-        Initialize a Cobertura report given a path to an XML file `xml_path`
-        that is in the Cobertura format.
+        Initialize a Cobertura report given a coverage report `report`. It can
+        be either a file object or the path to an XML file that is in the
+        Cobertura format.
 
-        The optional argument `base_path` can be provided to resolve the path
-        to the source code. If omitted, `base_path` will be set to
-        `os.path.dirname(xml_source)`. Additionally, `base_path` can be a path
-        to an archive zip or tarball) that contains the source code.
+        The optional argument `source` is the location of the source code
+        provided as a directory path or a file object zip archive containing
+        the source code.
+
+        The optional argument `source_prefix` will be used to lookup source
+        files if a zip archive is provided and will be prepended to filenames
+        found in the coverage report.
         """
-        if base_path is None:
-            base_path = os.path.dirname(xml_path)
-            self.filesystem = DirectoryFileSystem(base_path)
-        elif zipfile.is_zipfile(base_path):
-            self.filesystem = ZipFileSystem(base_path)
+        self.xml = ET.parse(report).getroot()
+
+        if source is None:
+            if isinstance(report, basestring):
+                # get the directory in which the coverage file lives
+                source = os.path.dirname(report)
+            else:
+                # fallback on using the current working directory
+                source = "%s/" % os.getcwd()
+            self.filesystem = DirectoryFileSystem(
+                source, source_prefix=source_prefix
+            )
+        elif zipfile.is_zipfile(source):
+            self.filesystem = ZipFileSystem(
+                source, source_prefix=source_prefix
+            )
         else:
-            self.filesystem = DirectoryFileSystem(base_path)
-
-        self.xml_path = xml_path
-        self.xml = ET.parse(xml_path).getroot()
+            self.filesystem = DirectoryFileSystem(
+                source, source_prefix=source_prefix
+            )
 
     @memoize
     def _get_class_element_by_filename(self, filename):